scanner/requirements.txt
ccxt
pandas
ta-lib
requests
mplfinance
psycopg2-binary
dotenv
